Create Resuable Steps
