+++ test/jest-component/features/steps/events/get-events-steps.ts
@@ -1,44 +1,17 @@
-import { Test, TestingModule } from '@nestjs/testing';
 import { StepDefinitions} from 'jest-cucumber';
 import * as request from 'supertest';
-import { AppModule } from '../../../../../src/app.module';
 import { HttpStatus, INestApplication } from '@nestjs/common';
 import { EventRepository } from '../../../../../src/infrastructure/db/repositories' ;
 import { futureEventMockDB, oldEventMockDB } from '../../../../mocks/db';
 import { matchers } from 'jest-json-schema';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 
 const addFormats = require('ajv-formats');
-const axios = require('axios');
-const SwaggerParser = require('swagger-parser');
-const Ajv = require('ajv');
+
 let app : INestApplication;
 let eventRepository: EventRepository;
 let schema;
 
-beforeEach(async () => {
-  const moduleFixture: TestingModule = await Test.createTestingModule({
-    imports: [AppModule],
-  }).compile();
-  app = moduleFixture.createNestApplication();
-  const swaggerConfig = new DocumentBuilder()
-  .setTitle('Esmorga API')
-  .setDescription('Swagger for Esmorga API.')
-  .setVersion('1.0')
-  .build();
-  const document = SwaggerModule.createDocument(app, swaggerConfig);
-  SwaggerModule.setup('swagger', app, document);
-  await app.init();
-  eventRepository = moduleFixture.get<EventRepository>(EventRepository);
-  const response = await request(app.getHttpServer()).get('/swagger-json');
-  const rawSchema = response.body
-  schema = await SwaggerParser.dereference(rawSchema);
-  
-});
 
-afterEach(async () => {
-  await app.close();
-});
 
 export const getEvents: StepDefinitions = ({ given, and, when, then}) => {
     let response
@@ -103,17 +76,6 @@ export const getEvents: StepDefinitions = ({ given, and, when, then}) => {
       }        
     });
     
-    and('the response should following swagger schema', () => {
-      const reference = schema.paths[path].get.responses[response.status].content['application/json'].schema
-      const validate = ajv.compile(reference);
-      const valid = validate(response.body);
-      if (!valid) {
-        console.error(validate.errors);
-      }
-      expect(valid).toBe(true);
-
-    });
-
     then(/^an error (\d+) in response should be returned$/, (error) => {
       if (error==500){
         expect(response.status).toBe(HttpStatus.INTERNAL_SERVER_ERROR);
