+++ test/jest-component/jest-cucumber-setup.ts
@@ -1,8 +1,15 @@
 
 import { loadFeatures, autoBindSteps } from 'jest-cucumber';
+import  './steps-config'
+const { directoryImport } = require('directory-import');
+const importedModules = directoryImport('./steps/');
 
-import { getEvents } from './events/get-events-steps';
+let Steps:any=[]
+Object.values(importedModules).forEach(file => {
+    Object.values(file).forEach(exported => {
+        Steps.push(exported)
+    })
+  });
 
 const features = loadFeatures('test/jest-component/features/*.feature')
-
-autoBindSteps(features, [ getEvents ]);
+autoBindSteps(features, Steps );
